fixed oscillator and other tests@
